Added loading screen for asthetics and product imaghe scraping
This patch allowed me to create a better loading screen and added basic loading animation to the load. its on a set timer and purely cosmetic but in the future i can run functions and pre-load everything during this time and it will be better then the app loading into a blacnk desktop.